fix: move JSSerde to a plain javascript file instead of a Svelte component
